[frontend] document instrumentation (#444)
