+++ docs/service_table.md
@@ -10,7 +10,7 @@ View [Service Graph](../README.md#architecture) to visualize request flows.
 | [currencyservice](../src/currencyservice/README.md)          | C++           | Converts one money amount to another currency. Uses real values fetched from European Central Bank. It's the highest QPS service.            |
 | [emailservice](./services/emailservice.md)                   | Ruby          | Sends users an order confirmation email (mock).                                                                                              |
 | [featureflagservice](./services/featureflagservice.md)       | Erlang/Elixir | CRUD feature flag service to demonstrate various scenarios like fault injection & how to emit telemetry from a feature flag reliant service. |
-| [frontend](../src/frontend/README.md)                        | JavaScript    | Exposes an HTTP server to serve the website. Does not require signup/login and generates session IDs for all users automatically.            |
+| [frontend](./services/frontend.md)                           | JavaScript    | Exposes an HTTP server to serve the website. Does not require signup/login and generates session IDs for all users automatically.            |
 | [loadgenerator](./services/loadgenerator.md)                 | Python/Locust | Continuously sends requests imitating realistic user shopping flows to the frontend.                                                         |
 | [paymentservice](./services/paymentservice.md)               | JavaScript    | Charges the given credit card info (mock) with the given amount and returns a transaction ID.                                                |
 | [productcatalogservice](./services/productcatalogservice.md) | Go            | Provides the list of products from a JSON file and ability to search products and get individual products.                                   |
